# docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Laravel API",
        "description": "API Documentation",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:80",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/cars/available": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Получить доступные автомобили",
                "description": "Возвращает список доступных автомобилей",
                "operationId": "eefb890d372509f396f3572e7d23ed34",
                "responses": {
                    "200": {
                        "description": "Список доступных автомобилей",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID автомобиля",
                                                "type": "integer"
                                            },
                                            "name": {
                                                "description": "Название автомобиля",
                                                "type": "string"
                                            },
                                            "configurations": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object"
                                                }
                                            },
                                            "prices": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Описание ошибки",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cars": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Получить список автомобилей",
                "operationId": "9f21c5e6d138e9d8aef915774c21ce5e",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Car"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "Создать новый автомобиль",
                "operationId": "4f29683227f73515ec772aaf3c21a4de",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Создано",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Car"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cars/{id}": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Получить один автомобиль",
                "operationId": "554c81d8b181f5a16af12f48ad9f07e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID автомобиля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Car"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Cars"
                ],
                "summary": "Обновить автомобиль",
                "operationId": "6d894d620e41826fd2239684671ecb33",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешное обновление",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Car"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cars"
                ],
                "summary": "Удалить автомобиль",
                "operationId": "967ce753e91bb20869b1b2162fa4ed7f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Автомобиль успешно удалён"
                    }
                }
            }
        },
        "/api/configurations": {
            "get": {
                "tags": [
                    "Configurations"
                ],
                "summary": "Получить список всех комплектаций",
                "operationId": "1c1d7ca866cd4811184b52ad650ab462",
                "responses": {
                    "200": {
                        "description": "Успешный ответ"
                    }
                }
            },
            "post": {
                "tags": [
                    "Configurations"
                ],
                "summary": "Создать новую комплектацию",
                "operationId": "142964c73cfc02f8b261bb90e932d7e6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConfigurationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Комплектация успешно создана"
                    }
                }
            }
        },
        "/api/configurations/{id}": {
            "get": {
                "tags": [
                    "Configurations"
                ],
                "summary": "Получить комплектацию по ID",
                "operationId": "4397310d0fe4da48fecc678172268da8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID комплектации",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ"
                    }
                }
            },
            "put": {
                "tags": [
                    "Configurations"
                ],
                "summary": "Обновить комплектацию по ID",
                "operationId": "725547eaeded1eeecbaf3d83c68f477f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID комплектации",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConfigurationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Комплектация успешно обновлена"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Configurations"
                ],
                "summary": "Удалить комплектацию по ID",
                "operationId": "5993a1b752619bf03c6c2c31cf0ce4f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID комплектации",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Комплектация успешно удалена"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CarRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Toyota Camry"
                    }
                },
                "type": "object"
            },
            "ConfigurationRequest": {
                "required": [
                    "car_id",
                    "name"
                ],
                "properties": {
                    "car_id": {
                        "description": "ID автомобиля",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Название комплектации",
                        "type": "string",
                        "example": "Comfort"
                    },
                    "option_ids": {
                        "description": "Массив ID опций",
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                },
                "type": "object"
            },
...
            "Car": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Toyota Camry"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}